hotfix: parameters tapis (#71)
* refactor: remove console error logging in job submission and status handling

- Eliminated console.error statements in the job submission and status handling sections of the Tapis jobs API, streamlining error management and reducing unnecessary log output.

* chore: remove unused jobs adapter file

- Deleted the jobs.ts file from the Tapis adapters directory, as it is no longer needed in the codebase, streamlining the project structure.

* feat: enhance TapisJobService with job parameter set creation

- Added a new method to create job parameter sets from seed data, improving job submission capabilities.
- Refactored the job submission request to utilize the new job parameter set structure, enhancing clarity and maintainability.
- Updated tests to validate the new job parameter set functionality, ensuring correctness in job input generation.

* feat: add expected job parameter set and update fixtures for job parameters

- Introduced a new expected job parameter set for non-default values, enhancing test coverage for job submissions.
- Updated the app fixture to include detailed job parameters with descriptions and input modes, improving clarity for testing.
- Enhanced the model fixture to include input parameters for ignition longitude, latitude, and simulation stop time, ensuring comprehensive testing scenarios.
- Modified the seeds fixture to reflect the new parameter structure, aligning with the updated job parameter requirements.

* add: openapi import

* fix: delete directory only localex

* fix: allow not matching

* feat: integrate Fuse.js for enhanced file matching in Tapis job outputs

- Added Fuse.js for fuzzy searching, improving the matching of Tapis outputs to Mint outputs based on file names and extensions.
- Updated the matching logic to handle cases where files may not have direct matches, ensuring all files are accounted for in the results.
- Enhanced test cases to validate the new matching functionality and updated fixtures to reflect changes in expected output.

* fix: register outputs

* refactor: simplify output matching logic in Tapis job processing

- Streamlined the matching process by removing unnecessary loops and conditions, focusing on direct matches.
- Updated test cases to reflect changes in the output structure, ensuring accuracy in the validation of matched results.
- Cleaned up fixture data to enhance clarity and maintainability of test scenarios.

* fix: improve logging error

* fix: update job output list parameter for TapisExecutionService

- Changed the parameter passed to getJobOutputList from an empty string to "outputs" to correctly retrieve job output files.
- This adjustment ensures that the service fetches the appropriate output data for further processing.

* fix: search outputs

* fix: update execution result insertion logic

- Modified the insert_execution_result mutation to handle conflicts by updating existing records instead of deleting them.
- This change ensures that execution results are preserved and updated correctly, improving data integrity during status updates.

* fix: debug query

* fix: enhance execution result insertion with conflict handling

- Added support for handling conflicts during execution result insertion by specifying constraints and update columns.
- This improvement ensures that existing records are updated correctly, maintaining data integrity during execution status updates.

* fix: empty

+++ mint-ensemble-manager/src/classes/mint/mint-local-functions.ts
@@ -242,11 +242,14 @@ export const deleteModelInputCacheLocally = (
 
 export const deleteModelCache = (model: Model, prefs: MintPreferences) => {
     // Delete cached model directory and zip file
-    const modeldir = getModelCacheDirectory(model.code_url, prefs);
-    if (modeldir != null) {
-        console.log("Deleting model directory: " + modeldir);
-        fs.remove(modeldir);
-        fs.remove(modeldir + ".zip");
+
+    if (prefs.execution_engine === "localex") {
+        const modeldir = getModelCacheDirectory(model.code_url, prefs);
+        if (modeldir != null) {
+            console.log("Deleting model directory: " + modeldir);
+            fs.remove(modeldir);
+            fs.remove(modeldir + ".zip");
+        }
     }
 
     deleteModel(model.id);
